Pulled out deduplication logic for use in all NodeTest implementations. Now return Seq instead of List from path evaluations
