+++ src/main/scala/com/atomist/tree/pathexpression/NamedNodeTest.scala
@@ -25,8 +25,7 @@ case class NamedNodeTest(name: String)
       Right(kids)
     case Descendant =>
       val possibleMatches: List[GraphNode] = Descendant.selfAndAllDescendants(tn).flatMap(n => findUnder(n)).toList
-      // TODO: not removing duplicates? NodesWithTag does. Pretty sure this should too. need to write a test
-      Right(possibleMatches)
+      Right(NodeTest.dedupe(possibleMatches))
   }
 
 }
